requirements.txt
pecan>=0.5.0
pymongo>=2.7.2
wsme
requests
watchdog
oslo.config
oslo.middleware
oslo.policy>=0.3.0
keystonemiddleware
PasteDeploy
influxdb==2.0.1
pika
python-surveilclient==0.5.1

